Fixed functional tests
Recent changes in TNS@space2 made functional tests hang.

As Eleftherious explained, for some astrocytes endfeet targets
are unnaturally away, which made attraction field degrade to 0;
in effect, synthesis could not reach endfeet targets.

To compensate for that, we set constant attraction field instead of
exponentially decreasing one.

Eleftherios added a check at TNS@space2 limiting the maximum number
of iterations, so that we get an Exception instead of hangup
in a situation similar to the one described above.

Change-Id: I98e201b5449bf3770e6ffc5bfd929fc76e2ce17f