Rewrite endfeet shrinking algorithms in numpy
Until now the target area distribution of the endfeet was sampled
from bio data and distributed to the grown endfeet in parrallel.
This constraining was not taking into account how big or small were
the grown areas in order to shrink them following the correct distribution.

The code was rewritten so that the grown areas are now possible to be
collected first, transformed with respect to the target distribution
following the correct probability mapping and be shrinked.

This gets rid of the cython code concerning the endfeet shrinking.

Changes Summary:

  - Gliovascular data is loaded at the app level
  - The parallel processing of the endfeet is removed. Everything was
    converted to numpy operations. In the end it is faster.
  - Distribution transformation of values implemented in area_mapping
  - Sorting of triangles and removal by travel time implemented in area_shrinking
  - The archngv.building.endfeet_reconstruction directory was collapsed
  - endfeet_reconstruction/detail/endfoot.py not needed anymore
  - endfeet_reconstruction/detail/area_fitting.py not needed
  - endfeet_reconstruction/detail/io.py migrated to exporters
  - fmm_growing moved one level up and was renamed to fast_marching_method.py
  - triangle grouping heleper functions added at endfeet_reconstruction/groups.py
  - Update fast_marching_method to be compatible with archngv-building 0.1.1
  - Improve non-overlapping search  when multiple endfeet converge to same mesh node
  - Fast marching method cutoff radius is a separate parameter

Change-Id: I120f4ace779918dc2fbd189863e51545703877e4