Make NGVConfig compatible with relative parent directory
The NGVConfig dictionary required an absolute path for the parent directory.
With this change the ngv_config.json that is created inside the circuit will
have "../" as parent directory which resolves depending on where the circuit
is located. That allows for easier portability of the circuit data. Of course
if an absolute path can be also provided.

Change-Id: I44dfb70e3653fb17361a50cdaccaed7fc0690c09